move global styles to styles folder
